+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+0.2.2
+-----
+- [isuftin@usgs.gov] - Updated to latest CIDA Auth version. 
+
+0.2.1
+-----
+
+- [isuftin@usgs.gov] - Tested as working with both Oracle and Postgres and the WSI Tomcat cookbook 
+- [isuftin@usgs.gov] - Added more testing. 
+
+
 0.2.0
 -----